Pin blinker to version <1.8. Fix for #329.
As PyleniumIO relies on selenium-wire (which is no longer alive) and this project again on blinker one would like to pin it to a working version. seleium-wire uses a private method of blinker which has been removed in the 1.8 release. As selenium-wire cannot be changed the only way is to pin blinker.